requirements.txt
google-genai
python-dotenv
groq

# for data extract from PDF
requests
pymupdf


pandas
openpyxl
feedparser



# summary

torch
torchvision
transformers

streamlit

numpy
neo4j
pyvis